+++ src/components/Tag/index.ts
@@ -1,2 +1,9 @@
 export { Tag } from "./Tag";
-export type { TagProps, TagColor, TagAppearance, TagShape } from "./Tag";
+export type {
+  TagProps,
+  TagColor,
+  TagStatusColor,
+  TagCategoryColor,
+  TagAppearance,
+  TagShape,
+} from "./Tag";